New Classes Show and Performance added to model package (see summary)
Upon realising the old class system didn't represent the data or model of how the theatre booking system should work, I deactivated the old classes ShowType and Concert and SeatsInterface and created the class Show and Performance. This should now model the theatre booking system better and be more in line with ERD and database requirements.